Change config files

+++ server/src/db/create.js
@@ -1,7 +1,13 @@
 const { sequelize } = require('./models');
+const { sequelizeInfo } = require('../config/index')
 
 async function main() {
-    await sequelize.sync({ force: true });
+    if(sequelizeInfo.syncType === 'FORCE')
+        await sequelize.sync({ force: true });
+    else if(sequelizeInfo.syncType === 'ALTER')
+        await sequelize.sync({ alter: true });
+    else
+        await sequelize.sync();
 }
 
 main()
